Add files via upload
Fixed incorrect site truncation for popups. Site and forecast are only truncated based on the LAST underscore, so that long site names are not incorrectly presented.

For example if site name is Dar_es_salaam, code used to write "Dar" as the site name and "Es" as the source. This is now fixed.